add conda configuration and Spark testing script
