--- config.py
@@ -1,7 +1,0 @@
-# LOGGER
-LOG_FORMAT = "%(asctime)s - %(name)s - %(levelname)s - %(funcName)s - line %(lineno)d - %(message)s"
-LOG_DIR = './logs'
-LOG_FILE = './logs/app.log'
-
-# DATA
-DATA_DIR = './data'
