+++ data_crawling/config.py
@@ -1,0 +1,7 @@
+# LOGGER
+LOG_FORMAT = "%(asctime)s - %(name)s - %(levelname)s - %(funcName)s - line %(lineno)d - %(message)s"
+LOG_DIR = '../logs'
+LOG_FILE = '../logs/crawler.log'
+
+# DATA
+DATA_DIR = '../data'
