--- requirements.txt
@@ -1,28 +1,0 @@
-async-generator==1.10
-attrs==22.1.0
-certifi==2022.9.24
-cffi==1.15.1
-charset-normalizer==2.1.1
-colorama==0.4.6
-exceptiongroup==1.0.0rc9
-h11==0.14.0
-idna==3.4
-outcome==1.2.0
-packaging==21.3
-pycparser==2.21
-pyparsing==3.0.9
-PySocks==1.7.1
-python-dotenv==0.21.0
-requests==2.28.1
-selenium==4.15
-sniffio==1.3.0
-sortedcontainers==2.4.0
-tqdm==4.64.1
-trio==0.22.0
-trio-websocket==0.9.2
-undetected-chromedriver==3.1.6
-urllib3==1.26.12
-webdriver-manager==3.8.4
-websockets==10.4
-wsproto==1.2.0
-tqdm==4.64.1
